// src/types/match.ts
export type Match = {
    _id: string;
    MatchType: string;
    MatchFormat: string;
    SchoolA: string;
    SchoolB: string;
    PlayerA1: string;
    PlayerA2: string;
    PlayerB1: string;
    PlayerB2: string;
    Game1A: string;
    Game2A: string;
    Game3A: string;
    Game1B: string;
    Game2B: string;
    Game3B: string;
    winner: string;
    owner: string;
};